react routing enabled
The main component is now again root.jsx Link component is styled and refactored into StyledLink element
Card on click navigate to ProjectInfo page . Logo and App name in navbar are links to Dashboard.
App.css is deletedn and font import shifted to root.css
