+++ app/src/components/StyledLink/StyledLink.css
@@ -1,0 +1,8 @@
+.link{
+    text-decoration: none;
+    color:black;
+}
+.link:focus, .link:hover, .link:visited, .link:link, .link:active {
+        text-decoration: none;
+        color:black;
+}
