+++ app/src/constants/routes.json
@@ -3,5 +3,6 @@
   "MOTD": "/motd",
   "LOCALIZATION": "/localization",
   "UNDOREDO": "/undoredo",
-  "CONTEXTMENU": "/contextmenu"
+  "CONTEXTMENU": "/contextmenu",
+  "PROJECTINFO":"/project/:projectName"
 }
