+++ app/src/core/root.css
@@ -1,3 +1,4 @@
+@import url('https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,400;0,700;1,300;1,400&family=Raleway:wght@300;400;600;700&display=swap');
 html {
   height: 100%;
   overflow-x: hidden;
