--- app/src/css/App.css
@@ -1,5 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Open+Sans:ital,wght@0,400;0,700;1,300;1,400&family=Raleway:wght@300;400;600;700&display=swap');
-.app{
-  background-color: white;  
-  width:100vw;
-}
