+++ app/src/index.jsx
@@ -4,7 +4,6 @@ import i18n from "I18n/i18n.config";
 import { I18nextProvider } from "react-i18next";
 import Root from "Core/root";
 import store, { history } from "Redux/store/store";
-import App from '../src/components/App';
 import './css/index.css'
 
 
@@ -12,7 +11,7 @@ ReactDOM.render(
   <div>
   <I18nextProvider i18n={i18n}>
     <Suspense fallback="loading">     
-       <App />
+    <Root store={store} history={history}></Root>
     </Suspense>
   </I18nextProvider>
   </div>,
